requirements.txt
tensorflow==2.16.1
torch==2.3.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0
opencv-python==4.10.0.84
Pillow==10.3.0
streamlit==1.36.0
tabpy==2.13.0

